Fix token reflection
Fixed the way tokens are reflected. Now it won't assume type is defined, and will make best effort to use token or type name.

Removed the duplicates check there because I was not sure what should it do (I can bring it back, just need to know how to integrate it correctly).

Resolves #25