refactor(common): change dependencies reflector signature
Refactor the `DependenciesReflector` in the common package to change its signature and return a different interface. This change was made to support both constructor parameters and properties. The motivation behind this change is to address a bug in the reflector where returning a Map does not allow dependencies to appear more than once.

Closes: https://github.com/automock/automock/issues/67

Please note that this is a breaking change.
BREAKING CHANGE: The `reflectDependencies` method of `DependenciesReflector` now returns a different interface to accommodate the support for both constructor parameters and properties.